+fix/FixUI, AddRotateCar

+++ Assets/Scripts/Bussiness/Enemy/Enemy.cs
@@ -6,5 +6,5 @@ public class Enemy : MonoBehaviour
 {
     public float health = 1f;
     public float shield = 1f;
-    public float scoreForPlayerWhenEnemyDead = 1f;
+    public int scoreForPlayerWhenEnemyDead = 1;
 }
